requirements.txt
# Discord
discord.py>=2.3.0

# Configuration
pyyaml>=6.0
python-dotenv>=1.0.0

# Database
aiosqlite>=0.19.0

# HTTP client
httpx>=0.25.0

# LLM providers
ollama>=0.3.0
openai>=1.0.0  # For OpenRouter compatibility
